LPS-179177 Extract CommerceCatalogChannelsScreenNavigationCategory methods to a CommerceCatalogChannelsScreenNavigationEntry component and extends CommerceCatalogChannelsScreenNavigationCategory in it
