+++ modules/apps/commerce/commerce-catalog-web/src/main/java/com/liferay/commerce/catalog/web/internal/frontend/taglib/servlet/taglib/CommerceCatalogChannelsScreenNavigationCategory.java
@@ -15,21 +15,12 @@
 package com.liferay.commerce.catalog.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.commerce.catalog.web.internal.constants.CommerceCatalogScreenNavigationConstants;
-import com.liferay.commerce.product.model.CommerceCatalog;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -38,15 +29,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Gianmarco Brunialti Masera
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceCatalogChannelsScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommerceCatalog> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -55,16 +42,11 @@ public class CommerceCatalogChannelsScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return getCategoryKey();
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, getCategoryKey());
+		return language.get(resourceBundle, getCategoryKey());
 	}
 
 	@Override
@@ -73,26 +55,7 @@ public class CommerceCatalogChannelsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_CATALOG_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, CommerceCatalog commerceCatalog) {
-		return false;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/commerce_catalog/channels.jsp");
-	}
-
 	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
+	protected Language language;
 
 }
